backend/requirements.txt
Django>=4.2.7,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
python-dotenv>=1.2.1
supabase>=2.10.0
PyJWT>=2.8.0
bcrypt>=4.0.0
Pillow>=10.1.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.3
gunicorn>=21.2.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.5
whitenoise>=6.6.0
gunicorn>=21.2.0


